Implements client list of repos
Sprint is incomplete. There are a lot of issues with asynchronocity and duplicate database entries. The Mongoose docs' recommended fix for the unique field only works half the time. Saving info to the database takes so long that React is re-rendering too quickly to show new data. Note to self: when should I return function calls that return promises, and when do I simply have the function called?